requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
PyMuPDF
python-docx
pandas
reportlab
tenacity
requests
pydantic
pdfplumber
openai
openrouter
fpdf